Add loading screen to back submission flow
